pref:hook_review_remove_button
